removing one vnr example config, simplifying the host script
