--- examples/bare-metal/vnr/config.xscope
@@ -1,22 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<!-- ======================================================= -->
-<!-- The 'ioMode' attribute on the xSCOPEconfig              -->
-<!-- element can take the following values:                  -->
-<!--   "none", "basic", "timed"                              -->
-<!--                                                         -->
-<!-- The 'type' attribute on Probe                           -->
-<!-- elements can take the following values:                 -->
-<!--   "STARTSTOP", "CONTINUOUS", "DISCRETE", "STATEMACHINE" -->
-<!--                                                         -->
-<!-- The 'datatype' attribute on Probe                       -->
-<!-- elements can take the following values:                 -->
-<!--   "NONE", "UINT", "INT", "FLOAT"                        -->
-<!-- ======================================================= -->
-
-<xSCOPEconfig ioMode="basic" enabled="true">
-    <Probe name="Value 1" type="CONTINUOUS" datatype="INT" units="Value" enabled="true"/>
-    <!-- For example: -->
-    <!-- <Probe name="Probe Name" type="CONTINUOUS" datatype="UINT" units="Value" enabled="true"/> -->
-    <!-- From the target code, call: xscope_int(PROBE_NAME, value); -->
-</xSCOPEconfig>
